requirements.txt
flask
python-dotenv
azure-ai-textanalytics
matplotlib